+++ packages/web-core/src/common/data_structures.ts
@@ -1,3 +1,4 @@
+import { InworldStatus as ProtoStatus } from '../../proto/ai/inworld/common/status.pb';
 import {
   CapabilitiesConfiguration,
   ClientConfiguration as ControlClientConfiguration,
@@ -12,6 +13,7 @@ import {
   SessionHistoryResponse,
 } from '../../proto/ai/inworld/packets/packets.pb';
 import { HistoryItem } from '../components/history';
+import { Character } from '../entities/character.entity';
 import { SessionContinuationProps } from '../entities/continuation/session_continuation.entity';
 import { AdditionalPhonemeInfo } from '../entities/packets/audio.entity';
 import { InworldPacket } from '../entities/packets/inworld_packet.entity';
@@ -150,11 +152,6 @@ export interface TriggerParameter {
   value: string;
 }
 
-export interface SendTriggerPacketParams {
-  parameters?: TriggerParameter[];
-  conversationId: string;
-}
-
 export enum InworldPacketType {
   UNKNOWN = 'UNKNOWN',
   TEXT = 'TEXT',
@@ -197,8 +194,27 @@ export interface PacketQueueItem {
   afterWriting: (packet: ProtoPacket) => void;
 }
 
+export enum MicrophoneMode {
+  UNSPECIFIED = 'UNSPECIFIED',
+  OPEN_MIC = 'OPEN_MIC',
+  EXPECT_AUDIO_END = 'EXPECT_AUDIO_END',
+}
+
 export interface SendPacketParams {
   conversationId: string;
+}
+
+export interface SendTriggerPacketParams extends SendPacketParams {
+  parameters?: TriggerParameter[];
+  character?: Character;
+}
+
+export interface SendAudioSessionStartPacketParams extends SendPacketParams {
+  mode?: MicrophoneMode;
+}
+
+export interface AudioSessionStartPacketParams {
+  mode?: MicrophoneMode;
 }
 
 export interface ConversationMapItem<
@@ -222,4 +238,10 @@ export interface ChangeSceneProps {
 export interface LoadedScene {
   sceneStatus: CurrentSceneStatus;
   sessionHistory?: SessionHistoryResponse;
+}
+
+export interface ProtoError {
+  message: string;
+  code: string | undefined;
+  details: ProtoStatus[] | undefined;
 }