+++ packages/web-core/src/index.ts
@@ -14,6 +14,7 @@ import {
   ConnectionConfig,
   Extension,
   HistoryChangedProps,
+  MicrophoneMode,
   StopAudioPlayback,
   TriggerParameter,
   User,
@@ -41,6 +42,12 @@ import {
   SessionContinuation,
   SessionContinuationProps,
 } from './entities/continuation/session_continuation.entity';
+import {
+  ErrorReconnectionType,
+  ErrorResourceType,
+  ErrorType,
+  InworldError,
+} from './entities/error.entity';
 import { DislikeType, Feedback } from './entities/feedback.entity';
 import {
   AdditionalPhonemeInfo,
@@ -90,6 +97,9 @@ export {
   EmotionEvent,
   EmotionStrength,
   EmotionStrengthCode,
+  ErrorReconnectionType,
+  ErrorResourceType,
+  ErrorType,
   Extension,
   Feedback,
   FeedbackDislikeProps,
@@ -104,8 +114,10 @@ export {
   HistoryItemTriggerEvent,
   InworldClient,
   InworldConnectionService,
+  InworldError,
   InworldPacket,
   InworldTriggers,
+  MicrophoneMode,
   PacketId,
   PreviousDialog,
   ProtoPacket,