+++ packages/web-core/src/services/inworld_connection.service.ts
@@ -1,5 +1,6 @@
 import { InworldPacket as ProtoPacket } from '../../proto/ai/inworld/packets/packets.pb';
 import {
+  AudioSessionStartPacketParams,
   AudioSessionState,
   CancelResponsesProps,
   ChangeSceneProps,
@@ -17,6 +18,7 @@ import { GrpcWebRtcLoopbackBiDiSession } from '../components/sound/grpc_web_rtc_
 import { InworldPlayer } from '../components/sound/inworld_player';
 import { InworldRecorder } from '../components/sound/inworld_recorder';
 import { Character } from '../entities/character.entity';
+import { InworldError } from '../entities/error.entity';
 import { InworldPacket } from '../entities/packets/inworld_packet.entity';
 import { EventFactory } from '../factories/event';
 import { characterHasValidFormat, sceneHasValidFormat } from '../guard/scene';
@@ -58,7 +60,7 @@ export class InworldConnectionService<
 
         if (!conversation) {
           this.connection.onError(
-            Error('No conversation is available to send audio.'),
+            new InworldError('No conversation is available to send audio.'),
           );
           return;
         }
@@ -224,7 +226,7 @@ export class InworldConnectionService<
     }
   }
 
-  async sendAudioSessionStart() {
+  async sendAudioSessionStart(params?: AudioSessionStartPacketParams) {
     if (this.connection.getAudioSessionAction() === AudioSessionState.START) {
       throw Error('Audio session is already started');
     }
@@ -233,7 +235,7 @@ export class InworldConnectionService<
 
     await this.ensureOneToOneConversation();
 
-    return this.oneToOneConversation.sendAudioSessionStart(true);
+    return this.oneToOneConversation.sendAudioSessionStart(params, true);
   }
 
   async sendAudioSessionEnd() {