+++ packages/web-core/CHANGELOG.md
@@ -1,8 +1,13 @@
 ## [Unreleased]
 
+### Added
+
+- Support Push-to-talk feature
+
 ### Changed
 
 - Support new runtime protocol session configuration
+- Expose Inworld error codes
 
 ## [2.6.0] - 2024-05-31